+++ tests/Admin/Feature/keiziban/DeleteMessageTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\keiziban;
+namespace Tests\Admin\Feature\keiziban;
 
 use App\Models\User;
 use App\Models\create_thread;
@@ -9,7 +9,7 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
-class RestoreMessageTest extends TestCase
+class DeleteMessageTest extends TestCase
 {
     use RefreshDatabase;
 
@@ -42,34 +42,34 @@ class RestoreMessageTest extends TestCase
         ]);
     }
 
-    public function test_not_login_get_access_restore_message()
+    public function test_not_login_get_access_delete_message()
     {
-        $response = $this->get('/jQuery.ajax/admin/restore_message');
+        $response = $this->get('/jQuery.ajax/admin/delete_message');
 
         $response->assertStatus(404);
     }
 
-    public function test_user_get_access_restore_message()
+    public function test_user_get_access_delete_message()
     {
         $response = $this
             ->actingAs($this->user)
-            ->get('/jQuery.ajax/admin/restore_message');
+            ->get('/jQuery.ajax/admin/delete_message');
 
         $response->assertStatus(404);
     }
 
-    public function test_admin_get_access_restore_message()
+    public function test_admin_get_access_delete_message()
     {
         $response = $this
             ->actingAs($this->admin)
-            ->get('/jQuery.ajax/admin/restore_message');
+            ->get('/jQuery.ajax/admin/delete_message');
 
         $response->assertStatus(404);
     }
 
-    public function test_not_login_post_access_restore_message()
+    public function test_not_login_post_access_delete_message()
     {
-        $response = $this->post('/jQuery.ajax/admin/restore_message', [
+        $response = $this->post('/jQuery.ajax/admin/delete_message', [
             'thread_id' => 'ThreadTestID',
             'message_id' => 1
         ]);
@@ -77,11 +77,11 @@ class RestoreMessageTest extends TestCase
         $response->assertStatus(404);
     }
 
-    public function test_user_post_access_restore_message()
+    public function test_user_post_access_delete_message()
     {
         $response = $this
             ->actingAs($this->user)
-            ->post('/jQuery.ajax/admin/restore_message', [
+            ->post('/jQuery.ajax/admin/delete_message', [
                 'thread_id' => 'ThreadTestID',
                 'message_id' => 1
             ]);
@@ -89,11 +89,11 @@ class RestoreMessageTest extends TestCase
         $response->assertStatus(404);
     }
 
-    public function test_admin_post_access_restore_message()
+    public function test_admin_post_access_delete_message()
     {
         $response = $this
             ->actingAs($this->admin)
-            ->post('/jQuery.ajax/admin/restore_message', [
+            ->post('/jQuery.ajax/admin/delete_message', [
                 'thread_id' => 'ThreadTestID',
                 'message_id' => 1
             ]);
