+++ tests/Admin/Feature/keiziban/SendMessageTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\hub;
+namespace Tests\Admin\Feature\keiziban;
 
 use App\Models\User;
 use App\Models\create_thread;
@@ -9,7 +9,7 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
-class DeleteThreadTest extends TestCase
+class SendMessageTest extends TestCase
 {
     use RefreshDatabase;
 
@@ -24,10 +24,10 @@ class DeleteThreadTest extends TestCase
 
         try {
             $create = new create_thread;
-            $create->create_thread('DeleteThreadTestID');
+            $create->create_thread('ThreadTestID');
             $create->insertTable(
-                'DeleteThreadTestName',
-                'DeleteThreadTestID',
+                'ThreadTestName',
+                'ThreadTestID',
                 'test@example.com'
             );
         } catch (QueryException $error) {
@@ -35,57 +35,60 @@ class DeleteThreadTest extends TestCase
         }
     }
 
-    public function test_not_login_get_access_admin_delete_thread()
+    public function test_not_login_get_access_send_message()
     {
-        $response = $this->get('jQuery.ajax/admin/delete_thread');
+        $response = $this->get('/jQuery.ajax/sendRow');
 
         $response->assertStatus(404);
     }
 
-    public function test_user_get_access_admin_delete_thread()
+    public function test_user_get_access_send_message()
     {
         $response = $this
             ->actingAs($this->user)
-            ->get('jQuery.ajax/admin/delete_thread');
+            ->get('/jQuery.ajax/sendRow');
 
         $response->assertStatus(404);
     }
 
-    public function test_admin_get_access_admin_delete_thread()
+    public function test_admin_get_access_send_message()
     {
         $response = $this
             ->actingAs($this->admin)
-            ->get('jQuery.ajax/admin/delete_thread');
+            ->get('/jQuery.ajax/sendRow');
 
         $response->assertStatus(404);
     }
 
-    public function test_not_login_post_access_admin_delete_thread()
+    public function test_not_login_post_access_send_message()
     {
-        $response = $this->post('jQuery.ajax/admin/delete_thread', [
-            'thread_id' => 'DeleteThreadTestID'
+        $response = $this->post('/jQuery.ajax/sendRow', [
+            'table' => 'ThreadTestID',
+            'message' => 'This is test comment!'
         ]);
 
-        $response->assertStatus(404);
+        $response->assertStatus(500);
     }
 
-    public function test_user_post_access_admin_delete_thread()
+    public function test_user_post_access_send_message()
     {
         $response = $this
             ->actingAs($this->user)
-            ->post('jQuery.ajax/admin/delete_thread', [
-                'thread_id' => 'DeleteThreadTestID'
+            ->post('/jQuery.ajax/sendRow', [
+                'table' => 'ThreadTestID',
+                'message' => 'This is test comment!'
             ]);
 
-        $response->assertStatus(404);
+        $response->assertStatus(200);
     }
 
-    public function test_admin_post_access_admin_delete_thread()
+    public function test_admin_post_access_send_message()
     {
         $response = $this
             ->actingAs($this->admin)
-            ->post('jQuery.ajax/admin/delete_thread', [
-                'thread_id' => 'DeleteThreadTestID'
+            ->post('/jQuery.ajax/sendRow', [
+                'table' => 'ThreadTestID',
+                'message' => 'This is test comment!'
             ]);
 
         $response->assertStatus(200);
