+++ tests/Admin/Feature/keiziban/ThreadLikeTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\keiziban;
+namespace Tests\Admin\Feature\keiziban;
 
 use App\Models\User;
 use App\Models\create_thread;
@@ -9,7 +9,7 @@ use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
-class SendMessageTest extends TestCase
+class ThreadLikeTest extends TestCase
 {
     use RefreshDatabase;
 
@@ -33,62 +33,73 @@ class SendMessageTest extends TestCase
         } catch (QueryException $error) {
             // nothing to do
         }
-    }
 
-    public function test_not_login_get_access_send_message()
-    {
-        $response = $this->get('/jQuery.ajax/sendRow');
-
-        $response->assertStatus(404);
-    }
-
-    public function test_user_get_access_send_message()
-    {
-        $response = $this
-            ->actingAs($this->user)
-            ->get('/jQuery.ajax/sendRow');
-
-        $response->assertStatus(404);
-    }
-
-    public function test_admin_get_access_send_message()
-    {
-        $response = $this
-            ->actingAs($this->admin)
-            ->get('/jQuery.ajax/sendRow');
-
-        $response->assertStatus(404);
-    }
-
-    public function test_not_login_post_access_send_message()
-    {
-        $response = $this->post('/jQuery.ajax/sendRow', [
-            'table' => 'ThreadTestID',
-            'message' => 'This is test comment!'
-        ]);
-
-        $response->assertStatus(500);
-    }
-
-    public function test_user_post_access_send_message()
-    {
-        $response = $this
+        $this
             ->actingAs($this->user)
             ->post('/jQuery.ajax/sendRow', [
                 'table' => 'ThreadTestID',
                 'message' => 'This is test comment!'
             ]);
+    }
+
+    public function test_not_login_get_access_thread_like()
+    {
+        $response = $this->get('/jQuery.ajax/like');
+
+        $response->assertStatus(404);
+    }
+
+    public function test_user_get_access_thread_like()
+    {
+        $response = $this
+            ->actingAs($this->user)
+            ->get('/jQuery.ajax/like');
+
+        $response->assertStatus(404);
+    }
+
+    public function test_admin_get_access_thread_like()
+    {
+        $response = $this
+            ->actingAs($this->admin)
+            ->get('/jQuery.ajax/like');
+
+        $response->assertStatus(404);
+    }
+
+    /*
+     * ログインしていない状態でのpost通信が防げない...
+     * この状態でpost通信は出来ないでしょうということでこのままで
+    */
+    public function test_not_login_post_access_thread_like()
+    {
+        $response = $this->post('/jQuery.ajax/like', [
+            'thread_id' => 'ThreadTestID',
+            'message_id' => 1
+        ]);
 
         $response->assertStatus(200);
     }
 
-    public function test_admin_post_access_send_message()
+    public function test_user_post_access_thread_like()
+    {
+        $response = $this
+            ->actingAs($this->user)
+            ->post('/jQuery.ajax/like', [
+                'thread_id' => 'ThreadTestID',
+                'message_id' => 1
+            ]);
+
+        $response->assertStatus(200);
+    }
+
+    public function test_admin_post_access_thread_like()
     {
         $response = $this
             ->actingAs($this->admin)
-            ->post('/jQuery.ajax/sendRow', [
-                'table' => 'ThreadTestID',
-                'message' => 'This is test comment!'
+            ->post('/jQuery.ajax/like', [
+                'thread_id' => 'ThreadTestID',
+                'message_id' => 1
             ]);
 
         $response->assertStatus(200);
